+++ src/pages/HomePage/components/Cart/Cart.tsx
@@ -6,7 +6,7 @@ import { CartItem } from './components/CartItem/CartItem';
 import { Delimiter } from '../../../../components/Delimiter/Delimiter';
 import { selectCartItems } from '../../../../store/reducers/cartSlice';
 import { useDispatch, useSelector } from 'react-redux';
-import { openAuthModal } from '../../../../store/reducers/signInOrSignUpSlice';
+import { openAuthModal } from '../../../../store/reducers/authModalSlice';
 import { useState } from 'react';
 
 interface CartProps {
